// client/src/interfaces/TextSettings.ts
interface TextSettings {
  fontFamily: string; // Check
  fontSize: string; // Check
  fontWeight: string; // Check
  fontStyle: string; // Check

  textOpacity: number; // Check
  lineHeight: number; // Check

  color: string; // Check
  backgroundColor: string; // Check

  letterSpacing: number; // Check
  wordSpacing: number; // Check
  textAlign: string;
}

export default TextSettings;